+++ src/plugins/core/parse/requestBodyHttpRegionParser.ts
@@ -12,7 +12,6 @@ export async function parseRequestBody(
     const next = lineReader.next();
     if (!next.done) {
       if (requestBody.rawBody.length > 0 || !utils.isStringEmpty(next.value.textLine)) {
-        requestBody.rawBody.push(parseLine(next.value.textLine));
         const symbols: Array<models.HttpSymbol> = [];
 
         if (!requestBody.symbol || requestBody.symbol.endLine !== next.value.line - 1) {
@@ -31,6 +30,24 @@ export async function parseRequestBody(
           requestBody.symbol.endLine = next.value.line;
           requestBody.symbol.endOffset = next.value.textLine.length;
           requestBody.symbol.children?.push?.(...utils.parseHandlebarsSymbols(next.value.textLine, next.value.line));
+        }
+
+        const fileImport = utils.parseFileImport(next.value.textLine);
+        if (fileImport) {
+          requestBody.rawBody.push(fileImport);
+          requestBody.symbol.children?.push(
+            new models.HttpSymbol({
+              name: 'filename',
+              description: fileImport.fileName,
+              kind: models.HttpSymbolKind.path,
+              startLine: next.value.line,
+              startOffset: next.value.textLine.indexOf(fileImport.fileName),
+              endLine: next.value.line,
+              endOffset: next.value.textLine.indexOf(fileImport.fileName) + fileImport.fileName.length,
+            })
+          );
+        } else {
+          requestBody.rawBody.push(next.value.textLine);
         }
 
         return {
@@ -52,8 +69,4 @@ export function getRequestBody(context: models.ParserContext) {
     context.data.request_body = result;
   }
   return result;
-}
-
-function parseLine(textLine: string) {
-  return utils.parseFileImport(textLine) || textLine;
 }